feat(core): version of axe without polyfills
